BIRNAT/requirement.txt
pytorch
scipy
scikit-image
opencv-python